matched styling to all pages and moved calendar heatmap to new component.
