+++ client/src/components/profile/genStats.js
@@ -7,27 +7,6 @@ import CalendarHeatmap from 'react-calendar-heatmap';
 
 
 export default function GenStats() {
-
-    const currentDate = new Date();
-    const startDate = new Date();
-    startDate.setMonth(startDate.getMonth() - 4);
-
-    const generateRandomValues = () => {
-        const values = [];
-        const currentDate = new Date();
-
-        for (let i = 0; i < 365; i++) {
-            const randomCount = Math.floor(Math.random() * 3) + 1;
-            const date = currentDate.toISOString().split('T')[0];
-            values.push({ date, count: randomCount });
-            currentDate.setDate(currentDate.getDate() - 1);
-        }
-
-        return values;
-    };
-
-    const randomValues = generateRandomValues();
-    console.log(randomValues)
     return (
         <div>
             <div className="flex flex-col items-center p-14 rounded-xl bg-white shadow-xl ">
@@ -58,24 +37,6 @@ export default function GenStats() {
                         </svg>
                         <span className="ml-1 text-sm text-gray-700">Languages Used: 3</span>
                     </div>
-                </div>
-            </div>
-
-            <div className="flex flex-col items-center p-14 rounded-xl mt-5 bg-slate-300 shadow-xl h-auto">
-
-                <CardTitle className="mb-8"> Submissions </CardTitle>
-                <div>
-                    <CalendarHeatmap
-                        startDate={startDate}
-                        endDate={currentDate}
-                        values={randomValues}
-                        classForValue={(value) => {
-                            if (!value) {
-                                return 'color-empty';
-                            }
-                            return `color-scale-${value.count}`;
-                        }}
-                    />
                 </div>
             </div>
         </div>
