+++ client/src/styles/globals.css
@@ -76,9 +76,10 @@
 }
 
 .cGradient{
-  background: rgb(31,41,55);
+    background: rgb(31,41,55);
     background: linear-gradient(180deg, rgba(31,41,55,1) 0%, rgba(255,255,255,1) 100%);
     border-top: 1px solid rgb(26, 35, 46);
+    width: 100%
 }
 p {
   margin: 20px 0px;
