+++ client/src/styles/home.css
@@ -1,3 +1,8 @@
+#navLogo{
+    width: 75px;
+    height: 60px;
+    margin-right: 10px;
+}
 .hero{
     background: rgb(31,41,55);
     background: linear-gradient(180deg, rgba(31,41,55,1) 0%, rgba(255,255,255,1) 100%);
